requirements.txt
fastapi
pydantic[email]
motor
python-decouple
uvicorn
